chore(sixlowpan): all 6lowpan logic into sixlowpan
